Restructure GitHub test reporter
